commet-remote-data-server/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyGithub==1.59.1
openai==1.12.0
python-dotenv==1.0.0
httpx==0.25.0
Flask-CORS==4.0.0
requests-oauthlib==1.3.1
requests>=2.31.0
